Massive refactoring undergone
I changed the changing model to be more object-based making the `replace` functions methods on a `String` type.
This allowed chaining methods and thus piping the output easily through them, I also added methods for common replacements
involving regexps, or the `strings.ReplaceAll` method, to handle errors and multiple cases in an encapsulated way.

The codebase should now be more legible and clear to understand.